// apps/desktop/src/types/generated/EdgeKindDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Mirrors [`EdgeKind`].
 */
export type EdgeKindDto = "assemblyRef" | "forceLoadAfter" | "forceLoadBefore" | "loadAfter" | "loadBefore" | "modDependency" | "findMod" | "ifModActive" | "patchTargetsDef" | "mayRequire" | "patchInjectedNode" | "assemblyVersionPrecedence" | "usesType" | "parentTemplate" | "patchRemovedNode" | "retextureAfterOwner" | "defOverrideAfterOrigin" | "patchSelectsInjectedNode" | "patchInvalidatesPredicate" | "patchRemovedNodeCosmetic" | "replaceDiscardsAddition";